Only attempt to install SendWP if not already installed

+++ includes/emails/functions.php
@@ -289,35 +289,45 @@ function wp_ajax_easy_digital_downloads_sendwp_remote_install_handler () {
 	include_once ABSPATH . 'wp-admin/includes/file.php';
 	include_once ABSPATH . 'wp-admin/includes/class-wp-upgrader.php';
 
-	/*
-	* Use the WordPress Plugins API to get the plugin download link.
-	*/
-	$api = plugins_api( 'plugin_information', array(
-		'slug' => $plugin_slug,
-	) );
+	$plugins = get_plugins();
 
-	if ( is_wp_error( $api ) ) {
-		ob_end_clean();
-		echo json_encode( array( 'error' => $api->get_error_message(), 'debug' => $api ) );
-		exit;
+	if( ! array_key_exists( 'sendwp/sendwp.php', $plugins ) ) {
+
+		/*
+		* Use the WordPress Plugins API to get the plugin download link.
+		*/
+		$api = plugins_api( 'plugin_information', array(
+			'slug' => $plugin_slug,
+		) );
+
+		if ( is_wp_error( $api ) ) {
+			ob_end_clean();
+			echo json_encode( array( 'error' => $api->get_error_message(), 'debug' => $api ) );
+			exit;
+		}
+
+		/*
+		* Use the AJAX Upgrader skin to quietly install the plugin.
+		*/
+		$upgrader = new Plugin_Upgrader( new WP_Ajax_Upgrader_Skin() );
+		$install = $upgrader->install( $api->download_link );
+		if ( is_wp_error( $install ) ) {
+			ob_end_clean();
+			echo json_encode( array( 'error' => $install->get_error_message(), 'debug' => $api ) );
+			exit;
+		}
+
+		/*
+		* Activate the plugin based on the results of the upgrader.
+		* @NOTE Assume this works, if the download works - otherwise there is a false positive if the plugin is already installed.
+		*/
+		$activated = activate_plugin( $upgrader->plugin_info() );
+
+	} else {
+
+		$activated = activate_plugin( 'sendwp/sendwp.php' );
+
 	}
-
-	/*
-	* Use the AJAX Upgrader skin to quietly install the plugin.
-	*/
-	$upgrader = new Plugin_Upgrader( new WP_Ajax_Upgrader_Skin() );
-	$install = $upgrader->install( $api->download_link );
-	if ( is_wp_error( $install ) ) {
-		ob_end_clean();
-		echo json_encode( array( 'error' => $install->get_error_message(), 'debug' => $api ) );
-		exit;
-	}
-
-	/*
-	* Activate the plugin based on the results of the upgrader.
-	* @NOTE Assume this works, if the download works - otherwise there is a false positive if the plugin is already installed.
-	*/
-	$activated = activate_plugin( $upgrader->plugin_info() );
 
 	/*
 	* Final check to see if SendWP is available.
